(maint) collapse error unless debug
Reduce the error message to just the message during normal operation. Expand
the message with more information during -d, /d, or any argument containing
debug.

Since the configuration variables may not be available, use primitive means of
seeing if the debug argument was specified. This means it cannot be combined
with other short args.