(config) remove useNugetForSources
Since we are no longer using nuget.exe, we would not be able to easily support
the ability to use the sources listed in nuget. This is actually a preferred
change as the two should be separated for configuration and the way chocolatey
supports it's configuration is a bit different than nuget.