separate parsing from validation
Previously there were issues when trying to -h commands if those commands were
doing validation in the handle additional parsing methods and not accounting
for whether help had been requested or not.

This fixes that by adding an additional handle_validation command to the
ICommand that is called to ensure the validation is correct, but only if help
has not been requested.