(log) PowershellService primitive hooks on logging
Implement primitive searching to determine the right logger to redirect
output to. This is primitive because it is line by line and some output
spans multiple lines. Only the first line will get flagged with color.

The future of the POSH service is to move to runspaces so no more work
should be put into making it work better.