(build) copy nuget.core over to build_output dir
Prior to compiling, copy nuget.core over to build_output directory to
attempt to resolve issue with finding NuGet in an existing location in
the packages directory. It exists but xbuild cannot seem to find it
there. This attempts to subvert the issue. NuGet.Core is the only
assembly with this issue, none of the others are left unresolved by
xbuild.