(GH-36) check for official public key
When choco.exe detects that it is not using the official
publickeytoken, throw an error requiring an explicit override (this is
already overridden in debug builds). When AllowUnofficialBuild flag is
set to true, log a very important warning so that folks can receive
clues that the state of their system may be compromised.