(GH-48) Ensure all Posh Errors are terminating
In Powershell, some errors are not subject to try catch blocks, which
is indeed confusing when trying to assign those particular errors to
try/catch blocks. Ensure that all errors coming back from a script are
terminating, because they will cause powershell to exit with a non-zero
status no matter what, which choco will fail the installation of the
package on. This makes things less confusing as to why that is
happening.