(maint) FileSystem combine paths unconditionally
Previously, if a section that was being combined to the  path starts
with `\` or `/`, it would override the first path completely. That means
you will get back none of the original path, only that section
forward. This is erroneous behavior.

As an example, `FileSystem.combine_paths("c:\\bob","\\joe")` would
return `\joe` instead of `c:\bob\joe`.

With this fix a section that is being combined to the original path
will no longer override the previous path.