(GH-47) Rollback when upgrade fails
Allow the user to determine if they want the previous package files to
be put back when an upgrade fails or is canceled. Handle state of
install, upgrade or uninstall that may have previous package folder
still around (usually due to unforeseen errors or user ctrl+c during
choco run).