(GH-54) Show help menu when unparsed options
If there are any unparsed switches/options after parsing is considered
complete, then show the help menu. This is introduced to address a
minor regression introduced by
https://github.com/chocolatey/choco/commit/51d59c82e01a8ca0ad6bef918789a58fd2faf441
, where the help menu would no longer be shown if an option was not
parsed. It was previously handling this by throwing an error that would
cause the help menu to show up. However b/c of the way the bundled
option parsing is done, coupled with multiple passes through the
parser, some options may be misinterpreted as bundled options and throw
errors, which was corrected in the linked commit.

Add additional behavior to detect unparsed args and show the help menu
instead of attempting to roll with it.