InstallLocation looks at ChocolateyInstall
Look first to see if there is an environment variable prior to using
relative location. Should work without issue with normal choco.exe
since those locations are the same when installed. Will need to figure
out how embedded DLL will interact with this properly if choco is NOT
installed on the machine.