(GH-67) Configure debug logging after local options
If debug (`-d`) is bundled with local options, it may not get picked up
when global options are parsed. If `config.Debug` is set once local
options are parsed, set the logging level again so that debug levels
are picked up and set against the appenders.

A few acceptable lines of debugging output will be missed but they are
related to writing out the command line that called choco with all of
the parameters and options.