(GH-81) Setting a config should Warn if no change
Previously if a config value was already set, there would be no output
letting the user know what happened. This is not helpful. When a config
value doesn't need to change, warn the user that the setting is already
set. Provide instructions on how to update a source (remove and add)