(GH-79) Determine when Feature set by user
Previously, a feature could be set by the system default or the user
explicitly setting the feature. When wanting to change feature flags
from false to true, there needs to be a way to determine if the user
explicitly set the flag so that the system doesn't automatically
override user intentions.

This will also warn when the user explicitly sets a value that is the
same as the default. Once explicitly set, the next time the setting is
attempted it will provide the no change message.