(GH-77) User is Administrator
Processes subject to UAC actually have the Administrators group
stripped out from the process, and will return false for any check
about being an administrator, including a check against the native
`CheckTokenMembership` or `UserIsAdmin`. Instead we need to perform a
not 100% answer on whether they are an admin based on if we have a split
token.

References
 -
http://www.davidmoore.info/blog/2011/06/20/how-to-check-if-the-current-user-is-an-administrator-even-if-uac-is-on/
-
http://blogs.msdn.com/b/cjacks/archive/2006/10/09/how-to-determine-if-a-user-is-a-member-of-the-administrators-group-with-uac-enabled-on-windows-vista.aspx

NOTE: from the latter (the original) -
```
Note that this technique detects if the token is split or not.  In the
vast majority of situations, this will determine whether the user is
running as an administrator. However, there are other user types with
advanced permissions which may generate a split token during an
interactive login (for example, the Network Configuration Operators
group). If you are using one of these advanced permission groups, this
technique will determine the elevation type, and not the presence (or
absence) of the administrator credentials.
```