(GH-97) Noop shows number of upgrades available
Previously the count of upgrades available was incorrectly just a count
of all packages, not just those available for upgrade. Noop with
upgrade is a pretty important function as it returns valuable
information. Show the number of packages, the number that are available
for upgrade, and the number of packages that have warnings (that are
not likely to be upgraded).