(GH-82) Set Modify for Installing User
With elevated permissions, choco can be installed. However sometimes
folks don't always run Chocolatey with elevated permissions. To combat
some of that and keep the permissions from being opened to all users,
add the installing user with explicit Modify permission to the
chocolatey install. This will allow them to perform basic functions and
not get errors when logging.