(GH-47) Rollback enhancements
If the primary rollback directory is not found, search for
directories that have the packagename and end with the rollback
suffix.

When moving the rollback to its previous name, ensure that you just
remove the rollback suffix so that the name of the folder is restored
as it was before.