(GH-98) Fix upgrade is a dependency of others
When attempting to upgrade packages, if they are a dependency of other
packages, it results in an error due to a packageManager uninstall
handler that is receiving an uninstall event without removing an older
package version. Do not add the eventhandler on upgrade so the package
manager can make the determination based purely on the upgrade scenario, not
based in a one off uninstallation attempt.