(specs) Adjust variables and unpack
For integration specs run against Release assembly, the paths are
already set to machine level when the application starts up. For
testing purposes, that needs to be readjusted back to the local
directory. Since many of these variables are readonly, update the
values with reflection.

Also ensure that the application unpacks itself locally.