(GH-120) ChocolateyPackageFolder env var wrong
In Powershell Choco, the chocolateyPackageFolder environment variable
would point to the directory the package was extracted to, e.g.
C:\ProgramData\chocolatey\lib\package.1.2.3 (relevant source lines:
https://github.com/chocolatey/chocolatey/blob/2d61a3b0958170ade13d8ad5a09a88af9e46bb65/src/functions/Run-ChocolateyPS1.ps1#L28
and
https://github.com/chocolatey/chocolatey/blob/2d61a3b0958170ade13d8ad5a09a88af9e46bb65/src/functions/Chocolatey-NuGet.ps1#L64).

In C# Choco, this variable is set to the path of the lib directory.
Set the variables to the proper directory for the package.