(GH-207) Set up variables for powershell scripts
Prior to calling the script, we need to set up variables so they are
available. To do that we redirect the call to a known powershell
script that will call the install/uninstall script. In this known
script we set up all of the variables appropriately so they are already
set when we call the actual script from the known script.