(GH-240) ChocolateyNugetCredentialProvider
When NuGet needs credentials for a source, it was using a default
credential provider that needs to use the same format as the NuGet
config file. This provides a way to pull those credentials from the
Chocolatey stored sources, use the passed in credentials, or prompt for
credentials if the user is interactive.