(GH-254) Ensure nupkg is removed on uninstall
NuGet will happily tell you that something has been uninstalled, even
if it didn't actually remove the nupkg from the folder. At the very
least we need an additional check to ensure the nupkg file is removed
in case NuGet has an issue removing the file.

Without this change the package may still be considered installed and
cause bad side effects when doing things like `cup all` or trying to
reinstall the package. This also resolves the requirement for cases
like this in using `--force` to remove the package.