(GH-212) Do not strip switch char from uninstall args
Previously, when building uninstall args, the uninstaller service was
splitting on ` /` and ` -`. While this was very accurate in finding the
first argument (the executable), it removed the switch characters from
the args. Use the same method to find the first argument to shift it
off, but then go back to the original uninstall string to gather the
arguments, removing the executable from that.

Additionally fix specs and improve clarity of the tests.