(GH-253) Fix Zip ChecksumType is ignored
Due to defaulting `$url64bit` to `$url`, `$checksumType` was converting
to the 64 bit checksumtype, even if it was empty. Which causes
checksumtype to default to md5, no matter what the checksumtype passed
in is.

Move the checksumType setting for x64 into the check if the urls for 32
bit and 64 bit are different. Additionally, ensure that the
checksumtype64 is not empty.