(GH-279) Resolve packages on first feed found
For aggregate repositories, packages are supposed to be resolved on the
first feed it is found on in priority order, no matter if it is the
latest version of that package or not. This is how it worked in 0.9.8.x
and below and how it should work until v1 when we can switch the
default. This fixes the behavior to be what is expected.