(GH-300) Allow upgrade to always install missing
The intent behind `choco upgrade git` issued against a machine is that
I want the latest version of git installed on that machine. Whether
that is upgrade or install doesn't really matter.

On a machine I issue the command `choco install git`, I just want to
ensure git is installed. If instead I issue the command `choco upgrade
git` on a machine, the end state I want for the machine is to have it
installed and up to date on the latest version of git. Whether git was
there before or not is not something I'm concerned with. The end state
is that I want git on the most up to date version based on Chocolatey.