(GH-14) Find path to executable
Provide a way to determine if a command / executable is on the file
system. It starts by getting the path extensions `PATHEXT`
environment variable and using those extensions combined with the
executable name (unless an extension is provided as part of the name)
and searching for the file in all path locations, starting at current
directory and moving out from there in the order things would be
searched for by the system.

If nothing is found, we just return the name that was passed into the method.