(GH-304) Auto Installer allow system cleanup
If the preceeding chocolatelyUninstall.ps1 uninstalls an application,
sometimes the system is still cleaning up files and registry keys.
Those keys are found by the auto installer and it decides to also move
forward with the uninstall. In cases like that, the auto uninstaller
should wait a few seconds before moving forward with checks.