(GH-305) Installer Type Validate Exit Code
When an uninstall finishes, the exit code of zero may not be the only
valid exit code. Allow the installer type to validate the exit code
against its valid uninstall exit codes. This will catch things like
with MSI with 0, 1641, and 3010 (all successful codes), and also 1605
and 1641 (for when the product was not installed, perhaps uninstalled
by a different means) as also valid since the end state is that the
application is no longer installed.