(GH-302) Fix: choco pin list fails
Sometimes when calling choco pin list, it fails to return a list of
valid items. This was due to "list" being added as a filter, so not all
locally installed items were being searched to see if they had a pin
for themselves.

Don't allow `config.Input` to be set or it will be passed along to list.