(GH-60) Fix: Pin errors on pkg folders w/versions
If the package folder has a version on it, it is considered a side by
side path, so the package manager may not be able to resolve the path
to the locally installed nupkg. When it isn't found by regular means,
attempt to find it by looking for it in side by side paths. If it is
not found there either, go ahead and error.