(GH-90) Fix install by nupkg file directly
When issuing a command like `choco install hg.3.3.nupkg`, choco tries to
find a package named "hg.3.3" (based on the passed in file name),
rather the reading the manifest of the nupkg and finding the actual
package name to install.

Fix the behavior to allow calling install on the nupkg file directly.