(GH-313) Snapshot detection of locked files
When a locked file is detected, use a special code for it instead of
file too big. When cleaning up the files during uninstall, provide
messages for the user to take action on those locked files as they will
be in the best place to determine whether they should stick around or
get removed.