(GH-205)(GH-194)(GH-290) Vendor NuGet.Core
Vendoring NuGet.Core allows us to take advantage of our own fixes to
the provider. There are things that have been wanted for awhile that
don't have a workaround to change the setting.

- GH-290 - packages no longer have a restriction on content folders
that produces the "External packages cannot depend on packages that
target projects". I'd like to call it a bug, and for Chocolatey it
certainly is.
- GH-194 - Fix minimum NuGet client version reports Chocolatey version
instead of NuGet's version, which is fine for packages on Chocolatey,
but the ones on NuGet usually have this filled in, which makes them
impossible to install with choco.
- GH-205 - Enhance the nuspec - first wave. Add the most oftenly
requested url for the nuspec, the packageSourceUrl and several other
urls plus add in replaces, provides and conflicts as placeholders for
now.