(GH-331) Use XDT transforms on configs
For Xml configuration files that are common with .NET, use Xml Document
Transformation (XDT) to change the file over time if the package maintainer
includes xdt files. These files are specified in a similar manner to
https://docs.nuget.org/Create/Configuration-File-and-Source-Code-Transformations
except we only support `*.install.xdt` files and they can be anywhere in the
package. They will apply to any configuration file with the same name.