(GH-293) Upgrade all --except
Allow passing a list of packages to skip on upgrade. There may be
times when you want to process upgrades for everything except for a
few applications that you may be currently running, like ConEmu.

Not being able to ignore certain packages on upgrade all is problematic,
especially when choco upgrades an application while a user is in the middle
of using it. This could get quite annoying, especially with something like
ConEmu - if you are using it to run upgarde all it will stop processing
choco and continue its upgrade. There are other situations where having the
ability to ignore certain packages would be very beneficial.