(GH-181)(GH-184) Short prompt/Prompt character
Right now, when asking a multiple choice question such as which version to
uninstall, there is no indication that a user input is expected. Add an input
indicator ">" to make this clearer.

Certain questions are spread out across multiple lines when they could
be on just one line as a short prompt that include the choices in parentheses.
For example, instead of

"Do you want to run the script?

1) yes
2) no
3) skip
>
"

it could simply be "Do you want to run the script? (Yes/No/Print): yes".

If the prompt allows for short answers, then provide hints that the answer
can be a single keystroke. For example
"Do you want to run the script? ([Y]es/[N]o/[P]rint): y"

In a short prompt, if there is a default answer, surround it with braces
as well. For example
"Do you want to run the script? ([[Y]es]/[N]o/[P]rint): _"