(GH-285) Fix download loop if connection lost
Previously, if the network connection is lost in the middle of
downloading a file, it will cause an infinite loop of error messages
similar to

~~~
Exception calling "Read" with "3" argument(s): "Unable to read data
from the transport connection:
~~~

Another unwelcome side effect is that the (corrupted) download file will
proceed to fill up all available disk space if left unchecked.

With this fix, throw immediately when this happens as there is no
recovery even if the network connection comes back up.