(GH-365) Exit Non-zero if newer version installed
If an older version of a package is requested and allow multiple
versions is not selected, choco will do nothing but exit with zero.
This gives the appearance that choco has installed that version for
automated tools when in fact it has done nothing. Instead it should
exit non-zero with an error.

Allow choco to attempt to install/upgrade to older versions with an
additional `--allow-downgrade` switch.