(GH-406) Passed powershell args retain quotes
When using powershell arguments ( `$packageParameters`) instead of the
environment variables (`$env:chocolateyPackageParameters`), be sure
that the arguments are able to retain any quotes that are passed in so
that the two variables match. Otherwise you get strange behavior.

Folks should not really be using the passed in arguments as they are
not part of the public API, but that doesn't mean they won't use them.