(GH-355) Write config to temp path first
If someone interrupts a write out for the config file, it can
permanently corrupt the config file and requires one to delete it and
let choco recreate it on the next run. Instead write the updates to the
config file to a new file and then copy it over the existing file. This
makes for a more atomic change that has very little chance of being
interrupted.