(GH-387) Validate before remove rollback folder
As a further enhancement for GH-341
(270ea940ee1bd48747f1968307f75c30023535d1),  ensure that
package names are not attempting to navigate out of the lib backup
directory. A specially crafted package name could cause choco to
attempt to delete folders it should not, therefore we need to restrict
it to the lib backup folder only.

If we find we are no longer in a subdirectory of the backup directory,
we should return immediately without attempting to delete anything.