(GH-347) Unpack choco files prior to nuget pack
Right now choco requires a functioning .NET 4+ framework install to
even install itself on a machine. During install It will attempt twice
to install .NET 4.0 if there isn't a version present, but it may run
into errors or require a reboot before it is ready to be used. Because
of this folks run into what looks like a missing file
`chocolateyInstaller.psm1` during install, when it's actually the former
reason. Without .NET 4.0, the files can't be unpacked from choco.

Unpacking those files in the nupkg will allow the errors to be more
visible on a non-functioning .NET framework install versus something
with a file missing.

This allows choco to be installed without the framework installed
properly, just not able to be used.