(GH-243) Ensure NuGet Proxy Override
NuGet originally set the ProxyCache internal which did not make it a
good candidate to use to set a custom proxy. Update NuGet.Core to one
that will allow that (our internal chocolatey-nuget). Add a method for
adding an override proxy server that will force NuGet to use the
override proxy if it exists.