(GH-452) Fix broken pin command
Choco pin list used to ensure that the output from the local `list_run`
command did not output on stdout. However a commit in GH-132 caused pin
to start outputting the list output as well as pin output. Fix this by
adding `config.QuietOutput = true;` around the call to List_run.

Also add integration spec scenarios to ensure this is not subject to
happen again.