(version) 0.9.10-beta1
(GH-460) Amended Template

Using 'choco new' to produce ChocolateyUninstall.ps1 for your package would
result in the uninstall block never being executed as the $key variable type
was wrong. Updated the script template to force the $key variable to always be
an array and therefore have a Count method that can be evaluated and the
uninstall block executed.

Closes #460