(GH-399) client certificate support
Extends authentication by adding options to sources and commands
to specify client certificate and its password, as well as
implementing new IClientCertificateProvider from choco/nuget-chocolatey
to lookup those options based on the Uri requested.
These options are complimentary and can be used in addition to
username/password assuming the server implements it.
+ unit tests fix to use lowercased %ComSpec% when looking for cmd.exe