(GH-507)(specs) Failing dependency version single tree
NuGet should walk a dependency tree and find the most appropriate
package by falling back down the dependency tree until it finds an
appropriate version that is suitable for all dependency sets and then
fail appropriately if no suitable version can be found.

Unfortunately when you use a Dependency Version Behavior of Highest,
NuGet just fails completely and refuses to work at all. I've been able
to verify this as high as the latest version of NuGet package manager
that is available with VS2013. I'm hopeful this is fixed in the v3.x of
NuGet.Core.