(GH-510) Fix: args for new fail with param=value
Some installers, MSIs in particular, specify silent arguments as
PROPERTY=value. Attempting to pass that in with
`choco new silentargs="/qn /norestart PROPERTY=value` will cause
choco to pass an empty string value in place of the passed arguments.

Fix the behavior to do the right thing when splitting the property
settings so that `=` can be passed as part of the arguments.