Adds WindowStyle, PinToTaskBar and RunAsAdmin capabilities.
WindowStyle parameter just exposes a standard WshShell.CreateShortcut property.  A link in the help section shows all possible values.

PinToTaskBar uses the shell.application object to call the shortcuts "Pin to Taskbar" shell verb.

RunAsAdmin - In order to enable the "Run as Admin" checkbox, this code reads the .LNK as a byte stream and flips a specific bit while writing a new copy.  It then replaces the original .LNK with the copy, similar to this example: http://poshcode.org/2513