(GH-8) Add System.Management.Automation binary
Add SMA and reference it appropriately. Through research it has been
found to take a dependency on the v1 version of SMA and set the
reference in the visual studio projects to not use a specific version
and not copy it local. This means it depends on being available in the
GAC, which should be if .NET Framework 4+ is installed.

The reference to v1 allows it to work across all versions of Windows
that have at least PowerShell v2 installed.