(GH-468) Log help to standard out, not standard error
Previously, help information was written out to stderr due to how NDesk
OptionSet provides the help output. This causes commands such as the following
to not page as expected:
* `choco list -help | more`
* `choco list -help | out-host -paging`

Similarly `choco -help | out-file HelpDetails.txt` does not deliver all of the
help output to the file.

While there is a workaround of using `2>&1` (or `*>&1`) is burdensome for the
very people who need help in the first place.

Help information should instead be provided in stdout and not in stderr. This
allows us to avoid these issues entirely.