(GH-476) Verbose log file operations / checksum
Any file system operation is probably better served as a verbose log as
it could be quite a bit of information being passed around.

For checksumming files, it is definitely a verbose set of information -
definitely not something we want to overload users with unless they opt
into seeing the information.