(GH-493) Reorder Search Result Information
The order provided doesn't flow well for what folks are looking
for when they are looking at verbose output about a package.
Reorder it in the way that provides the most helpful information
about the package first, then the software and finally the
description as it can take up quite a bit of space.