(GH-584) Hook for loading licensed posh extensions
If the license is valid, provide a hook for loading the posh
extensions. They will check on the license themselves and fail hard, so
it is best to provide this hook so that choco will continue to work
with only warnings if the license expires and the user hasn't done
anything.