(GH-585) Force should allow downgrade
Previously, force would do everything except for allow downgrades to
occur. There is a special switch for downgrades (`--allow-downgrades`)
and it was required to be passed to allow for a downgrade to occur.
When a user issues the force command they expect that it should
override anything necessary in the install/upgrade, including a
downgrade, because they are forcing it (force means overriding the
behavior of the system). Allow force to override the check for
downgrades.