(GH-584) Reinstall should use available package
The information contained in the available package may have pertinent
information that the installed package does not have, such as whether
it is approved or not, or whether there is a cached download available.
Use the available package for performing the reinstall versus the
existing package. This also may fix other issues, like when a package
has been corrected while it is still under moderation and/or while a
maintainer is repeatedly testing a package.