(GH-479) Remove NuGet cache of package
Chocolatey doesn't use the NuGet cache at all, but NuGet always caches
the package in the NuGet cache (and there is not a way to shut it off
with NuGet calls unfortunately). This can cause issues when later using
NuGet and trying to install a package from NuGet.org that is different,
but carries the same name. The cached file will be served to the user
causing hard to detect issues. Therefore we need to clean up after
NuGet as there is no way to tell it NOT to cache the nupkg during
install/upgrade.