(GH-503) Credential request should mask password
when a source and/or a proxy does not have the credentials stored in
the chocolatey.config file or when they are incorrect, choco will
prompt for valid credentials. When requesting proxy/network password
for a source, the password should be masked on the screen so that the
password is not able to be seen by other folks.