(GH-466) Credential cache validates against base url
First atttempt to validate against the exact path for the source Uri,
but when that fails due to subpaths in the source uri string,
also attempt to validate credentials against the host Uri as well.

In most cases this should be fine, however there could be a very slight
security issue if the host domain is tracking calls on the credentials
when the sub path is owned by someone different, such as in shared
hosting where the domain is shared and the subpaths are different for
everyone. The chances of this are extremely slight given that most
hosting for NuGet Servers own the virtual site or are used internally.
In most cases those host domains probably already validate your user
name/password in other places. It is worth noting though. Only use
hosting solutions you trust.