(GH-631) Remove Get-BinRoot, Add Get-ToolsLocation
Get-ToolsLocation is the successor to Get-BinRoot. It sets an alias to
the old method and provides the deprecation message when this method is
called with the older alias.

Get-ToolsLocation is a much more aptly named value, and the environment
variable surrounding it makes sense upon sight.  If either of the older
environment variables are found, they are used to update this value and
then those older environment variables are removed.

This also resolves GH-628, the tools location can be the root of the
drive again.