(GH647)(GH-374) Update shimgem
Shimgen took a dependency on something that was super heavyweight (it
increased the size by almost 8 times!) and was later found to be
completely unnecessary. Switched it over to using built-in libraries
and the size is back to what it was but still able to provide the same
features as the larger binary did.