(GH-258) File detector use exe
After testing with the c++ dll, it appears the DIE utility is only
available as an x86, which makes it difficult to use without bringing
in COM. As it is simpler to just use the executable it comes with
instead, switch over to using that. Set it up with an ini file to
configure it to find the detector database in the proper location as
well.