(GH-663) Search by Id Starts With
As an additional enhancement to searching by id, allow searching where
the id starts with the filter. This adds to tab completion.

The filters should be applied prior to searching by all versions gets
returned, so move the filters up.

Ensure the search is case-insensitive if it is local. A remote search
is case insensitive if the remote server supports it.