(GH-572) Fix: The handle is invalid - Output Redirection
When running PowerShell but redirecting the output and a package
attempts to write to a filestream, it runs into an age old PowerShell
bug "The OS handle's position is not what FileStream expected. Do not
use a handle simultaneously in one FileStream and in Win32 code or
another FileStream. This may cause data loss."

Usually this bug is known to affect PowerShell v2, but when running
PowerShell with System.Management.Automation, it appears to affect more
versions (it has been seen in PowerShell v3 and v4).

Running this in newer versions of PowerShell has been tested not to
cause any adverse affects, so it is presumed to be safe.

The fix comes from
http://www.leeholmes.com/blog/2008/07/30/workaround-the-os-handles-position-is-not-what-filestream-expected/
plus information in the comments that actually allows it to work.