(GH-674)(GH-672) Color Overrides / Fix Write-Host
Write-Host was double printing, but it appears that it was due to
calling Write, and then an empty WriteLine, both of which the
PoshHostUserInterface were calling a log entry which writes out each as
a line. Instead log just to the file and print with `Console.Write` and
`Console.WriteLine`.  Using `Console` calls solves the issue of not
seeing the colorization overrides, so those will be displayed properly.