(GH-572) Console Adapter Redirected Output
When output is being redirected, `IConsole` needs to be able to handle
that as most of `System.Console` throws errors on methods and
properties that would normally be available.

Detect if output is being redirected by checking the StdHandle against
FileType.

Additionally, handle certain calls to find out about console related
information.