(GH-572) Fix: The handle is invalid
When redirecting output, if anything attempts to write to the file
stream, it causes this error. The error is related to using Console
related properties and values in `PoshHostRawUserInterface`. The fix is
to use the `IConsole` adapter instead of calling `System.Console`
directly, as `IConsole` already has the ability to handle redirected
output.