(GH-586) Ignore invalid options/switches by default
Much like the way PowerShell ignores switches it doesn't understand,
allow doing that with swtiches and options that choco doesn't
understand. This allows passing things that would be created in future
versions and not having older versions break when attempting to use
those switches. Also allow for the feature to be turned off.