(GH-696) Approved/Failing statuses on install
Warn folks on install/upgrade if a package already has known failures.
This same information is provided in search results and on the site,
which could help them spend more time researching prior to attempting
installs of bad packages.

It's also helpful to know if this was an approved package that they
were installing/upgrading.