(GH-702) Fully qualify shutdown.exe
Previously, choco would use the first shutdown that it came across in
the path, which may not be the shutdown that it should be looking to
use. This could cause issues.

With this change, choco will fully qualify the path to the
shutdown.exe. If a user is running under SysWOW, the shutdown they will
find will be in the redirected path, but it is still valid to handle
the canceling of any shutdown/restart operations in progress.