(GH-512) Allow configuring package exit codes
Package exit codes provide the ability for the exit code of an
underlying installer to be passed through to Chocolatey's exit code.
This value is the last non-zero exit code from a package or dependency
that Chocolatey is currently installing. This change, while
very enabling, may be surprising for some folks and they may wish to
shut this feature off.

As this feature may need to be shut off, provide it as both a feature
and switches to turn it on/off per install, upgrade, and uninstall. Add
the switches to tab expansion as well.