(maint) Provide awareness of pro / business
Once every ten runs or  so, randomly mention the professional and
business versions. This will help bring awareness for users that won't
be aware in any other way.

When a user is using install directory override through the install
arguments, provide awareness that there is a feature that will save
them the trouble of spending extra time determining the install
directory themselves.